refactor: remove context-menu demo module
